requirements.txt
jsonrpclib==0.2.1
Pillow==10.3.0
PyYAML==6.0.1
